# openapi/devices-api.yml
openapi: 3.0.3
info:
  title: Devices API
  description: REST API for managing devices
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /devices:
    post:
      tags:
        - Devices
      summary: Create a new device
      operationId: createdDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRequest'
      responses:
        '201':
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'

    get:
      tags:
        - Devices
      summary: Fetch All devices with pagination
      operationId: getDevices
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
            example: createdAt,desc
        - name: state
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DeviceState'
        - name: brand
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Paginated devices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDeviceResponse'

  /devices/{id}:
    get:
      tags:
        - Devices
      summary: Fetch a device by id
      operationId: getDeviceById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Device found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
    put:
      tags:
        - Devices
      summary: Update device
      operationId: updateDevice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdateRequest'
      responses:
        '200':
          description: Device update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
    patch:
      tags:
        - Devices
      summary: Partially update a device
      operationId: patchDevice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevicePatchRequest'
      responses:
        '200':
          description: Device update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'

    delete:
      tags:
        - Devices
      summary: Delete a device
      operationId: deleteDevice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Device deleted

components:
  schemas:
    DeviceRequest:
      type: object
      required:
        - name
        - brand
        - state
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        brand:
          type: string
          minLength: 1
          maxLength: 100
        state:
          $ref: '#/components/schemas/DeviceState'
    DeviceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        brand:
          type: string
        state:
          $ref: '#/components/schemas/DeviceState'
        createdAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
          description: Used for optimistic locking during updates
    DeviceState:
      type: string
      x-java-type: com.devices.devices_service.domain.DeviceState
    PageDeviceResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/DeviceResponse'
        page:
          type: integer
        size:
          type: integer
        totalElement:
          type: integer
          format: int64
        totalPages:
          type: integer
    DeviceUpdateRequest:
      type: object
      required:
        - name
        - brand
        - state
        - version
      properties:
       name:
        type: string
       brand:
        type: string
       state:
          $ref: '#/components/schemas/DeviceState'
       version:
         type: integer
         format: int64
    DevicePatchRequest:
      type: object
      required:
        - version
      properties:
        name:
          type: string
        brand:
          type: string
        state:
          $ref: '#/components/schemas/DeviceState'
        version:
          type: integer
          format: int64